requirements.txt
beautifulsoup4==4.11.1
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.2
cycler==0.11.0
filelock==3.8.0
fonttools==4.37.4
fvcore==0.1.5.post20220512
gdown==4.4.0
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.7
iopath==0.1.9
kiwisolver==1.4.4
lmdb==1.3.0
matplotlib==3.6.1
munkres==1.1.4
numpy==1.22.3
packaging==21.3
pandas==1.5.1
patsy==0.5.3
Pillow==9.2.0
pip==22.3
portalocker==2.5.1
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
# pytorch3d==0.6.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
scipy==1.6.2
seaborn==0.11.2
setuptools==65.5.0
six==1.16.0
soupsieve==2.3.2.post1
statsmodels==0.13.2
tabulate==0.9.0
termcolor==2.0.1
toolz==0.11.2
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.1
typing_extensions==4.4.0
unicodedata2==14.0.0
urllib3==1.26.11
wheel==0.37.1
yacs==0.1.8
